feat: add example & refine code.
