+++ .gitignore
@@ -1,3 +1,4 @@
 /target
 .idea
 .idea/workspace.xml
+.obsidian
