+++ twopc/src/ot.rs
@@ -2,13 +2,10 @@
 
 use std::io::Cursor;
 use num_bigint::BigUint;
-use rand::thread_rng;
 use rrsa_lib::key::{Key, KeyPair};
-use crate::circuit::Circuit::{Gate, Input};
-use crate::circuit::{GateInfo, PartyInput};
 use crate::error::OTError;
 use crate::error::OTError::{DecryptionFailed, EncryptionFailed, InsufficientEncryptedPassword, LengthNotMatch};
-use crate::garbled_circuit::{garble_circuit, EncryptedValue};
+use crate::garbled_circuit::EncryptedValue;
 use crate::password::{Password, PasswordPair};
 
 // R = Random number which is agreed by both parties.
@@ -87,11 +84,11 @@ impl NKeysList {
 
         let mut res = Vec::new();
 
-        for i in 0..encrypted_value_list.len() {
+        for (i, item) in encrypted_value_list.iter().enumerate()  {
             let pos = self.positions[i];
-            let encrypted_value = encrypted_value_list[i][pos as usize].clone().0;
+            let encrypted_value = item[pos as usize].clone().0;
             let mut passwords_byte = vec![];
-            self.private_keys[i].decode(&mut Cursor::new(encrypted_value), &mut passwords_byte).map_err(|e| DecryptionFailed(e))?;
+            self.private_keys[i].decode(&mut Cursor::new(encrypted_value), &mut passwords_byte).map_err(DecryptionFailed)?;
 
             let password = Password::deserialize(<&[u8; 33]>::try_from(passwords_byte.as_slice()).unwrap());
             res.push(password);
@@ -129,15 +126,20 @@ pub fn ot_encryption_passwords(passwords_b: &[PasswordPair], nkeys_list: &[NKey]
     }
     let mut res = Vec::new();
     for i in 0..passwords_b.len() {
+
         if nkeys_list[i].0.len() != N as usize || passwords_b[i].pair.len() != N as usize{
             return Err(LengthNotMatch);
         }
+
+        if !nkeys_list[i].verify() {
+            return Err(OTError::InvalidNKeys)
+        }
         let mut enc_gate_value = vec![EncryptedValue::default(); N as usize];
-        for j in 0..nkeys_list[i].0.len() {
+        for (j, enc_gate_value) in enc_gate_value.iter_mut().enumerate().take(nkeys_list[i].0.len()) {
             let serialized_password = passwords_b[i].pair[j].serialize();
             let mut encrypted_value = Vec::new();
-            nkeys_list[i].0[j].encode(&mut Cursor::new(serialized_password), &mut encrypted_value).map_err(|e| EncryptionFailed(e))?;
-            enc_gate_value[j] = EncryptedValue(encrypted_value);
+            nkeys_list[i].0[j].encode(&mut Cursor::new(serialized_password), &mut encrypted_value).map_err(EncryptionFailed)?;
+            *enc_gate_value = EncryptedValue(encrypted_value);
         }
         res.push(enc_gate_value.clone().try_into().expect("Failed to convert to [Passsword; N as usize]"));
 
@@ -146,66 +148,76 @@ pub fn ot_encryption_passwords(passwords_b: &[PasswordPair], nkeys_list: &[NKey]
     Ok(res)
 }
 
-#[test]
-pub fn test_gen_key() {
-    let secret_value = vec![0, 1, 1];
-    let nkeylist = NKeysList::new(&secret_value);
-    for nkey in nkeylist.nkeys {
-        assert!(nkey.verify());
+mod test {
+    use std::io::Cursor;
+    use rand::thread_rng;
+    use rrsa_lib::key::KeyPair;
+    use crate::circuit::Circuit::{Gate, Input};
+    use crate::circuit::{GateInfo, PartyInput};
+    use crate::garbled_circuit::garble_circuit;
+    use crate::ot::{ot_encryption_passwords, NKeysList};
+    use crate::password::Password;
+
+    #[test]
+    pub fn test_gen_key() {
+        let secret_value = vec![0, 1, 1];
+        let nkeylist = NKeysList::new(&secret_value);
+        for nkey in nkeylist.nkeys {
+            assert!(nkey.verify());
+        }
     }
-}
 
-#[test]
-pub fn test_encryption_and_decryption_passwords() {
-    let mut rng = thread_rng();
-    let pass = Password::new(&mut rng, 0);
-    let serialized_pass = pass.serialize();
-    println!("pass: {:?}", serialized_pass);
+    #[test]
+    pub fn test_encryption_and_decryption_passwords() {
+        let mut rng = thread_rng();
+        let pass = Password::new(&mut rng, 0);
+        let serialized_pass = pass.serialize();
+        println!("pass: {:?}", serialized_pass);
 
-    let key_pair = KeyPair::generate(Some(512), true, true, true);
-    let pub_key = key_pair.public_key;
-    let priv_key = key_pair.private_key;
-    let mut encrypted_value = Vec::new();
+        let key_pair = KeyPair::generate(Some(512), true, true, true);
+        let pub_key = key_pair.public_key;
+        let priv_key = key_pair.private_key;
+        let mut encrypted_value = Vec::new();
 
-    pub_key.encode(&mut Cursor::new(serialized_pass), &mut encrypted_value).unwrap();
-    println!("{:?}", encrypted_value);
+        pub_key.encode(&mut Cursor::new(serialized_pass), &mut encrypted_value).unwrap();
+        println!("{:?}", encrypted_value);
 
-    let mut decrypted_value = Vec::new();
-    priv_key.decode(&mut Cursor::new(encrypted_value), &mut decrypted_value).unwrap();
+        let mut decrypted_value = Vec::new();
+        priv_key.decode(&mut Cursor::new(encrypted_value), &mut decrypted_value).unwrap();
 
-    println!("{:?}", decrypted_value);
+        println!("{:?}", decrypted_value);
 
-    assert_eq!(decrypted_value, serialized_pass);
+        assert_eq!(decrypted_value, serialized_pass);
 
-}
+    }
 
-#[test]
-pub fn test_ot_flow() {
-    let circuit = Gate(
-        GateInfo(0b1000),
-        Box::new(Gate(GateInfo(0b1001), Box::new(Input(PartyInput::A(0))), Box::new(Input(PartyInput::B(0))))),
-        Box::new(Gate(GateInfo(0b1001), Box::new(Input(PartyInput::A(1))), Box::new(Input(PartyInput::B(1))))),
-    );
-    let mut rng = thread_rng();
-    let (_, passwords_a, passwords_b, _) = garble_circuit(&mut rng, &circuit);
+    #[test]
+    pub fn test_ot_flow() {
+        let circuit = Gate(
+            GateInfo(0b1000),
+            Box::new(Gate(GateInfo(0b1001), Box::new(Input(PartyInput::A(0))), Box::new(Input(PartyInput::B(0))))),
+            Box::new(Gate(GateInfo(0b1001), Box::new(Input(PartyInput::A(1))), Box::new(Input(PartyInput::B(1))))),
+        );
+        let mut rng = thread_rng();
+        let (_, _, passwords_b, _) = garble_circuit(&mut rng, &circuit);
 
-    let a_value = vec![0_u8, 1_u8];
-    let b_value = vec![0_u8, 0_u8];
+        let b_value = vec![0_u8, 0_u8];
 
-    // Bob generates public keys
-    let nkey_list = NKeysList::new(&b_value);
+        // Bob generates public keys
+        let nkey_list = NKeysList::new(&b_value);
 
-    // Bob sends public keys for Alice, and Alice encrypts passwords with these keys
-    let res_encryption_passwords = ot_encryption_passwords(&passwords_b, &nkey_list.nkeys);
-    assert!(res_encryption_passwords.is_ok());
-    let encryption_passwords = res_encryption_passwords.unwrap();
+        // Bob sends public keys for Alice, and Alice encrypts passwords with these keys
+        let res_encryption_passwords = ot_encryption_passwords(&passwords_b, &nkey_list.nkeys);
+        assert!(res_encryption_passwords.is_ok());
+        let encryption_passwords = res_encryption_passwords.unwrap();
 
-    // Bob encrypts all needed passwords
-    let res_passwords = nkey_list.decrypt(encryption_passwords);
-    assert!(res_passwords.is_ok());
-    let passwords = res_passwords.unwrap();
+        // Bob encrypts all needed passwords
+        let res_passwords = nkey_list.decrypt(encryption_passwords);
+        assert!(res_passwords.is_ok());
+        let passwords = res_passwords.unwrap();
 
-    // Check the encrypted passwords are correct.
-    assert_eq!(passwords[0], passwords_b[0].pair[0]);
-    assert_eq!(passwords[1], passwords_b[1].pair[0]);
+        // Check the encrypted passwords are correct.
+        assert_eq!(passwords[0], passwords_b[0].pair[0]);
+        assert_eq!(passwords[1], passwords_b[1].pair[0]);
+    }
 }
